django_backend/requirements.txt

Django==4.2.10
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-dotenv==1.0.1
Pillow==10.2.0
# For production deployment
gunicorn==21.2.0
